flake8-requirements.txt
flake8
flake8-annotations-complexity
flake8-bandit
flake8-bugbear
flake8-builtins
flake8-class-attributes-order
flake8-commas
flake8-comprehensions

flake8-functions
mypy-extensions

flake8-import-order
flake8-print
flake8-quotes
flake8-simplify
flake8-variables-names
pep8-naming
django-stubs
django-stubs-ext